to prevent circular import move configs to config.py and setup code to a new file called setup.py
